pyproject.toml
[tool.poetry]
name = "carbon-aware-scheduler"
version = "0.1.0"
description = "A carbon aware scheduler based on Electricity Maps's forecasts."
authors = ["ElectricityMaps"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-cloud-pubsub = "^2.27.3"
google-cloud-storage = "^3.0.0"
openapi-client = {path = "./carbon-aware-api-client", develop = true}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
